\name{QCA.Notation}
\alias{QCA.Notation}
\title{
  Notations used in QCA3 package
}
\description{
  This page describes some notations used in QCA3 package
}
\details{
  -9: don't case case.
  
  *: boolean operator of and.

  +: boolean operator of or.
  
  cond\{v\}: condition has a value of v.
  
  A/B: condition A happend before B.
  
  NA: missing data.
}
