\name{Stokke}
\alias{Stokke}
\docType{data}
\title{
Effectiveness of international regimes for resource management
}
\description{
This data set can be used to demonstrate the usage of simplifying assumptions.
}
\usage{data(Stokke)}
\format{
  A data frame with 9 observations on the following 6 variables.
  \describe{
    \item{\code{advice}}{whether the shamers can substantiate their criticism by reference to explicit advice by the regime's scientific body.}
    \item{\code{commitment}}{commitment is a broad term, referring not only to legally binding conservation rules.}
    \item{\code{shadow}}{the shadow of the future.}
    \item{\code{inconvenience}}{inconvenience of the behavioral change.}
    \item{\code{reverberation}}{if a strong domestic constituency exits that reverberates the internationalcriticism.}
    \item{\code{success}}{success of shaming.}
  }
}
\references{
Stokke, O. S.2007. 'Qualitative comparative analysis, shaming, and international regime effectiveness'. Journal of Business Research,60 (5): 501-511.
}
\examples{
## refer to example of constrReduce
}
\keyword{datasets}
